+++ zcash_voting/src/prelude.rs
@@ -79,17 +79,19 @@ pub use crate::types::{
     MAX_VOTE_OPTIONS, MIN_PROPOSAL_ID, MIN_VOTE_OPTIONS,
 };
 pub use crate::vote::{
-    commit as commit_vote, commit_batch, parse_recovery,
+    commit as commit_vote, commit_batch, parse_recovery, persist_prepared_commit,
+    persist_prepared_commit_batch, prepare_commit, prepare_commit_batch,
     record_submission as record_vote_submission, record_vc_position,
     recover_commit as recover_vote_commit, recover_signed_commitments, recovery_bundle,
     serialize_recovery, submission as vote_submission, validate_draft_vote, validate_draft_votes,
-    CommittedVote, DraftVote, SignedVoteCommitment, SignedVoteCommitments, VanWitness, VoteCommit,
-    VoteCommitStage, VoteRecoveryBundle, VoteSigner, VoteSubmission,
+    CommittedVote, DraftVote, PreparedVoteCommit, PreparedVoteCommitments, SignedVoteCommitment,
+    SignedVoteCommitments, VanWitness, VoteCommit, VoteCommitBatch, VoteCommitStage,
+    VoteRecoveryBundle, VoteSigner, VoteSubmission,
 };
-pub use crate::warm_proving_caches;
 pub use crate::wire::{
     DelegationSubmissionWire, VoteCommitmentWire, VoteShareWire, VotingHotkeyTargetV1,
 };
+pub use crate::{warm_proving_caches, warm_zkp2_proving_cache};
 
 pub use crate::precompute::delegation_pir;